DLRA_Vlasov

Documentation for DLRA_Vlasov.

DLRA_Vlasov.A_dot

DLRA_Vlasov.A_dot

DLRA_Vlasov.K_dot!

DLRA_Vlasov.L_dot!

DLRA_Vlasov.S_dot!

DLRA_Vlasov.calc_E!

DLRA_Vlasov.calc_c1!

DLRA_Vlasov.calc_c2!

DLRA_Vlasov.calc_d1!

DLRA_Vlasov.calc_d2!

DLRA_Vlasov.A_dot — Method

A_dot(X,S,V,p,t)

Right-hand side of the Vlasov-Poisson equation, using a low-rank representation.

Arguments

X0::AbstractArray: the left basis matrix with orthonormal columns.

S0::AbstractArray: the coefficient matrix of the initial data.

V0::AbstractArray: the right basis matrix with orthonormal columns.

p::NamedTuple: the parameters.

t::AbstractFloat: the time of evaluation.

DLRA_Vlasov.A_dot — Method

A_dot(Y,p,t)

Right-hand side of the Vlasov-Poisson equation.

Arguments

Y0::AbstractArray: the initial data.

p::NamedTuple: the parameters.

t::AbstractFloat: the time of evaluation.

DLRA_Vlasov.K_dot! — Method

K_dot!(K̇,K0,p,t)

Right-hand side of the DLRA K-step for Vlasov-Poisson equation.

Arguments

K̇::AbstractArray: the matrix to be mutated.

K0::AbstractArray: the initial data.

p::NamedTuple: the parameters.

t::AbstractFloat: the time of evaluation.

DLRA_Vlasov.L_dot! — Method

L_dot!(L̇,L0,p,t)

Right-hand side of the DLRA L-step for Vlasov-Poisson equation.

Arguments

L̇::AbstractArray: the matrix to be mutated.

L0::AbstractArray: the initial data.

p::NamedTuple: the parameters.

t::AbstractFloat: the time of evaluation.

DLRA_Vlasov.S_dot! — Method

S_dot!(Ṡ,S0,p,t)

Right-hand side of the DLRA S-step for Vlasov-Poisson equation.

Arguments

Ṡ::AbstractArray: the matrix to be mutated.

S0::AbstractArray: the initial data.

p::NamedTuple: the parameters.

t::AbstractFloat: the time of evaluation.

DLRA_Vlasov.calc_E! — Method

calc_E(X0,S0,V0,k,h)

Solve the Poisson equation for the inrotational electric field.

Arguments

X0::AbstractArray: the left basis matrix with orthonormal columns.

S0::AbstractArray: the coefficient matrix of the initial data.

V0::AbstractArray: the right basis matrix with orthonormal columns.

h::AbstractFloat: the step-size of the velocity discretization.

dx::Tuple: a triplet containing a vector to be mutated, a frequency vector, and a pre-planned FFT object.

DLRA_Vlasov.calc_c1! — Method

calc_c1!(c1,v,V0)

Compute the c^1 coefficient matrix for the Vlasov-Poisson equation.

Arguments

c1::AbstractArray: the matrix to be mutated.

v::AbstractArray: the velocity vector.

V0::AbstractArray: the initial data.

DLRA_Vlasov.calc_c2! — Method

calc_c2!(c1,dv,V0,dvV)

Compute the c^2 coefficient matrix for the Vlasov-Poisson equation.

Arguments

c2::AbstractArray: the matrix to be mutated.

dv::Tuple: a triplet containing a vector to be mutated, a frequency vector, and a pre-planned FFT object.

V0::AbstractArray: the initial data.

dvV::AbstractArray: the matrix to be mutated for the spectral computation of the derivative.

DLRA_Vlasov.calc_d1! — Method

calc_d1!(d1,E,X0)

Compute the d^1 coefficient matrix for the Vlasov-Poisson equation.

Arguments

d1::AbstractArray: the matrix to be mutated.

E::AbstractArray: the electric field vector.

X0::AbstractArray: the initial data.

DLRA_Vlasov.calc_d2! — Method

calc_d2!(d2,dx,X0,dxX)

Compute the d^2 coefficient matrix for the Vlasov-Poisson equation.

Arguments

d2::AbstractArray: the matrix to be mutated.

dx::Tuple: a triplet containing a vector to be mutated, a frequency vector, and a pre-planned FFT object.

X0::AbstractArray: the initial data.

dxX::AbstractArray: the matrix to be mutated for the spectral computation of the derivative.