DLRA_Vlasov

Documentation for DLRA_Vlasov.

DLRA_Vlasov.A_dot

DLRA_Vlasov.A_dot

DLRA_Vlasov.BUG_step!

DLRA_Vlasov.K_dot!

DLRA_Vlasov.L_dot!

DLRA_Vlasov.S_dot!

DLRA_Vlasov.calc_E!

DLRA_Vlasov.calc_c1!

DLRA_Vlasov.calc_c2!

DLRA_Vlasov.calc_d1!

DLRA_Vlasov.calc_d2!

DLRA_Vlasov.get_sol

DLRA_Vlasov.mBUG_step!

DLRA_Vlasov.A_dot — Method

A_dot(X0,S0,V0,p,t)

Right-hand side of the Vlasov-Poisson equation, using a low-rank representation.

Arguments

X0::AbstractArray: the left basis matrix with orthonormal columns.

S0::AbstractArray: the coefficient matrix of the initial data.

V0::AbstractArray: the right basis matrix with orthonormal columns.

p::NamedTuple: the parameters.

t::AbstractFloat: the time of evaluation.

Description

The Vlasov-Poisson equation is given by the following expression,

\[\partial_t f(t,x,v) = -v \cdot \nabla_x f(t,x,v) + E(f)(x) \cdot \nabla_v f(t,x,v),\]

but now further does a low-rank representation of $f$, i.e.

\[ f(t,x,v) = ∑_{i,j=1}^r X_i(t,x) S_{ij}(t) V_{j}(t,v).\]

DLRA_Vlasov.A_dot — Method

A_dot(Y0,p,t)

Right-hand side of the Vlasov-Poisson equation.

Arguments

Y0::AbstractArray: the initial data.

p::NamedTuple: the parameters.

t::AbstractFloat: the time of evaluation.

Description

The Vlasov-Poisson equation is given by the following expression,

\[ \partial_t f(t,x,v) = -v \cdot \nabla_x f(t,x,v) + E(x) \cdot \nabla_v f(t,x,v).\]

DLRA_Vlasov.BUG_step! — Method

BUG_step!(X0,V0,S0,K0,L0,tmpx,tmpv,c1,c2,d1,d2,p,r,rmax,augment)

Perform a complete step with the rank-adaptive BUG integrator (Ceruti et al., 2022) for the Vlasov Poisson equation, and update the basis and coefficient matrices.

DLRA_Vlasov.K_dot! — Method

K_dot!(K̇,K0,p,t)

Right-hand side of the DLRA K-step for Vlasov-Poisson equation.

Arguments

K̇::AbstractArray: the matrix to be mutated.

K0::AbstractArray: the initial data.

p::NamedTuple: the parameters.

t::AbstractFloat: the time of evaluation.

Description

Following (Einkemmer and Lubich, 2018), the differential equation for the K-step is,

\[ \dot{K}_j = \sum_{l=1}^r \left( - c_{jl}^1 ⋅ \nabla_x K_l (t,x) + c_{jl}^2 ⋅ E(K)(t,x) K_l(t,x) \right).\]

DLRA_Vlasov.L_dot! — Method

L_dot!(L̇,L0,p,t)

Right-hand side of the DLRA L-step for Vlasov-Poisson equation.

Arguments

L̇::AbstractArray: the matrix to be mutated.

L0::AbstractArray: the initial data.

p::NamedTuple: the parameters.

t::AbstractFloat: the time of evaluation.

Description

Following (Einkemmer and Lubich, 2018), the differential equation for the L-step is,

\[ \dot{L}_i(t,v) = \sum_{k=1}^r \left(- (d_{ik}^2 \cdot v) L_k(t,v) + d_{ik}^1 \cdot \nabla_v L_k(t,v) \right).\]

DLRA_Vlasov.S_dot! — Method

S_dot!(Ṡ,S0,p,t)

Right-hand side of the DLRA S-step for Vlasov-Poisson equation.

Arguments

Ṡ::AbstractArray: the matrix to be mutated.

S0::AbstractArray: the initial data.

p::NamedTuple: the parameters.

t::AbstractFloat: the time of evaluation.

Following (Einkemmer and Lubich, 2018), the differential equation for the S-step is,

\[ \dot{S}_{ij}(t) = \sum_{k,l=1}^r \left(- c_{jl}^1 \cdot d_{ik}^2 + c_{jl}^2 \cdot d_{ik}^1 \right) S_{kl}(t).\]

DLRA_Vlasov.calc_E! — Method

calc_E!(E,X0,S0,V0,h,dx)

Solve the Poisson equation for the inrotational electric field.

Arguments

E::AbstractArray: the electric field vector to be mutated.

X0::AbstractArray: the left basis matrix with orthonormal columns.

S0::AbstractArray: the coefficient matrix of the initial data.

V0::AbstractArray: the right basis matrix with orthonormal columns.

h::AbstractFloat: the step-size of the velocity discretization.

dx::Tuple: a triplet containing a vector to be mutated, a frequency vector, and a pre-planned FFT object.

Description

The electric field in the case of the Vlasov-Poisson equation, satisfies the following equations,

\[ \nabla {E}(f)(x) = 1 - \int_{\Omega_v} f(t,x,v) \,\mathrm{d} v, \qquad \nabla \times {E}(f)(x) = 0.\]

DLRA_Vlasov.calc_c1! — Method

calc_c1!(c1,v,V0)

Compute the $c^1$ coefficient matrix for the Vlasov-Poisson equation.

Arguments

c1::AbstractArray: the matrix to be mutated.

v::AbstractArray: the velocity vector.

V0::AbstractArray: the initial data.

Description

Following (Einkemmer and Lubich, 2018), the $c^1$ coefficient matrix is given by,

\[ c^1_{jl} \coloneqq \int_{\Omega_v} v V_j V_l \,\mathrm{d} v.\]

DLRA_Vlasov.calc_c2! — Method

calc_c2!(c1,dv,V0,dvV)

Compute the $c^2$ coefficient matrix for the Vlasov-Poisson equation.

Arguments

c2::AbstractArray: the matrix to be mutated.

dv::Tuple: a triplet containing a vector to be mutated, a frequency vector, and a pre-planned FFT object.

V0::AbstractArray: the initial data.

dvV::AbstractArray: the matrix to be mutated for the spectral computation of the derivative.

Description

Following (Einkemmer and Lubich, 2018), the $c^2$ coefficient matrix is given by,

\[ c^2_{jl} \coloneqq \int_{\Omega_v} V_j (\nabla_v V_l) \,\mathrm{d} v.\]

DLRA_Vlasov.calc_d1! — Method

calc_d1!(d1,E,X0)

Compute the $d^1$ coefficient matrix for the Vlasov-Poisson equation.

Arguments

d1::AbstractArray: the matrix to be mutated.

E::AbstractArray: the electric field vector.

X0::AbstractArray: the initial data.

Description

Following (Einkemmer and Lubich, 2018), the $d^1$ coefficient matrix is given by,

\[ d^1_{ik} \coloneqq \int_{\Omega_x} X_i E X_k \,\mathrm{d} x.\]

DLRA_Vlasov.calc_d2! — Method

calc_d2!(d2,dx,X0,dxX)

Compute the $d^2$ coefficient matrix for the Vlasov-Poisson equation.

Arguments

d2::AbstractArray: the matrix to be mutated.

dx::Tuple: a triplet containing a vector to be mutated, a frequency vector, and a pre-planned FFT object.

X0::AbstractArray: the initial data.

dxX::AbstractArray: the matrix to be mutated for the spectral computation of the derivative.Description

Following (Einkemmer and Lubich, 2018), the $d^2$ coefficient matrix is given by,

\[ d_{ik}^2 \coloneqq \int_{\Omega_x} X_i(\nabla_x X_k) \,\mathrm{d} x.\]

DLRA_Vlasov.get_sol — Method

get_sol(N,r,τ,T,tol,flag,integrator,augment)

Compute and store a solution of a specific initial condition specified by flag, given by integrator at time T, given a discretization by N spatially and by τ in time.

DLRA_Vlasov.mBUG_step! — Method

mBUG_step!(X0,V0,S0,K0,L0,tmpx,tmpv,c1,c2,d1,d2,p,r,rmax,ic,augment)

Perform a complete step with the rank-adaptive BUG integrator (Ceruti et al., 2024) for the Vlasov Poisson equation, and update the basis and coefficient matrices.

References

Ceruti, G.; Einkemmer, L.; Kusch, J. and Lubich, C. (2024). A robust second-order low-rank BUG integrator based on the midpoint rule. BIT Numerical Mathematics 64.

Ceruti, G.; Kusch, J. and Lubich, C. (2022). A rank-adaptive robust integrator for dynamical low-rank approximation. BIT Numerical Mathematics 62, 1149–1174.

Einkemmer, L. and Lubich, C. (2018). A Low-Rank Projector-Splitting Integrator for the Vlasov–Poisson Equation. SIAM Journal on Scientific Computing 40, B1330–B1360.